+++ CHANGELOG.md
@@ -25,4 +25,8 @@ Development now switches to the Dev branch.
 2022/12/15 -- v0.1.2
 Minor bug fixes
 
-
+2022/12/19 -- v0.1.3
+ * Major refactoring
+ * The stopOptoStim method implements a rampdown
+ * Fix bugs that were causing waveforms to not be what were expected
+ * Laser is disabled for a fixed number of ms when location switching. Before it was 1 sample.